backend/requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.4

# Redis
redis==5.0.1

# Celery for background tasks
celery==5.3.6
kombu==5.3.4

# Email (SMTP)
aiosmtplib==3.0.1
email-validator==2.1.0

# HTTP client
httpx==0.26.0
aiohttp==3.9.1

# Content extraction
trafilatura==1.6.3
readability-lxml==0.8.1
beautifulsoup4==4.12.3
lxml==5.1.0

# RSS parsing
feedparser==6.0.10

# ML/Embeddings
openai==1.9.0
anthropic==0.18.1
tiktoken==0.5.2
numpy==1.26.3

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.0
tenacity==8.2.3

# Security
passlib[bcrypt]==1.7.4
PyJWT==2.8.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==24.1.0
ruff==0.1.13
mypy==1.8.0
